refactor to use an object
