+++ src/main/java/org/opengis/cite/cdb10/metadataAndVersioning/DefaultsXmlStructureTests.java
@@ -20,95 +20,26 @@ import java.util.*;
 public class DefaultsXmlStructureTests extends CommonFixture {
     @Test
     public void verifyDefaultsXmlFileExists() {
-        Assert.assertTrue(Files.exists(Paths.get(path, "Metadata", "Defaults.xml")),
-                "Metadata directory should contain Defaults.xml file.");
+        new DefaultsXml(path);
     }
 
     @Test
     public void verifyDefaultsXmlAgainstSchema() throws IOException, SAXException {
-        File xmlFile = Paths.get(path, "Metadata", "Defaults.xml").toFile();
-        File xsdFile = Paths.get(path, "Metadata", "Schema", "Defaults.xsd").toFile();
-
-        SchemaValidatorErrorHandler errorHandler = XMLUtils.validateXmlFileIsValid(xmlFile, xsdFile);
-
-        if (!errorHandler.noErrors()) {
-            Assert.fail(xmlFile.getName() + " does not contain valid XML. Errors: " + errorHandler.getMessages());
-        }
+        new DefaultsXml(path).verifyXmlAgainstSchema();
     }
 
     @Test
     public void verifyDefaultsXmlElementR_W_TypeHasValidValues() {
-        NodeList nodeList = XMLUtils.getNodeList("//R_W_Type", Paths.get(path, "Metadata", "Defaults.xml"));
-
-        ArrayList<String> values = new ArrayList<>();
-        List<String> VALID_VALUES = Arrays.asList("W", "R");
-
-        for (int i = 0; i < nodeList.getLength(); i++) {
-            Node currentItem = nodeList.item(i);
-            values.add(currentItem.getTextContent());
-        }
-
-        for (String value : values) {
-            Assert.assertTrue(VALID_VALUES.contains(value),
-                    String.format("Defaults.xml element R_W_Type should have a value of R or W. Value '%s' is not valid.", value));
-        }
+        new DefaultsXml(path).verifyElementR_W_TypeHasValidValues();
     }
 
     @Test
     public void verifyDefaultsXmlNameIsUniqueForEachDataset() {
-        for (String datasetValue : getDatasetValues()) {
-            assertNamesAreUnique(datasetValue);
-        }
+        new DefaultsXml(path).verifyNameIsUniqueForEachDataset();
     }
 
     @Test
     public void verifyDefaultsXmlElementTypeHasValidValue() {
-        NodeList nodeList = XMLUtils.getNodeList("//Type", Paths.get(path, "Metadata", "Defaults.xml"));
-
-        ArrayList<String> values = new ArrayList<>();
-        List<String> VALID_VALUES = Arrays.asList("string", "integer", "float");
-
-        for (int i = 0; i < nodeList.getLength(); i++) {
-            Node currentItem = nodeList.item(i);
-            values.add(currentItem.getTextContent());
-        }
-
-        for (String value : values) {
-            Assert.assertTrue(VALID_VALUES.contains(value),
-                    String.format("Defaults.xml element Type should have a value of " +
-                            "'float', 'integer' or 'string'. Value '%s' is not valid.", value));
-        }
+        new DefaultsXml(path).verifyElementTypeHasValidValue();
     }
-
-    private HashSet<String> getDatasetValues() {
-        NodeList datasetNodes = XMLUtils.getNodeList("//Dataset", Paths.get(path, "Metadata", "Defaults.xml"));
-
-        HashSet<String> datasetValues = new HashSet<>();
-
-        for (int i = 0; i < datasetNodes.getLength(); i++) {
-            Node currentItem = datasetNodes.item(i);
-            datasetValues.add(currentItem.getTextContent());
-        }
-        return datasetValues;
-    }
-
-    private ArrayList<String> collectNameNodesWithDatasetValue(String datasetValue) {
-        NodeList nameNodes = XMLUtils.getNodeList("//Default_Value[Dataset/text() = \"" + datasetValue + "\"]/Name", Paths.get(path, "Metadata", "Defaults.xml"));
-
-        ArrayList<String> names = new ArrayList<>();
-        for (int i = 0; i < nameNodes.getLength(); i++) {
-            names.add(nameNodes.item(i).getTextContent());
-        }
-        return names;
-    }
-
-    private void assertNamesAreUnique(String datasetValue) {
-        ArrayList<String> names = collectNameNodesWithDatasetValue(datasetValue);
-
-        for (String name : names) {
-            Assert.assertEquals(Collections.frequency(names, name), 1,
-                    String.format("Defaults.xml element Name should be unique under each Dataset. '%s' is not unique.", name));
-        }
-    }
-
 }
